Merge in count to magnitude and count to flux density conversions.
These should eventually be promoted to gphoton2 proper but for now
it’s convenient to have them in glcat.

The `Band` type is now a dataclass (with only two instances) that
carries useful ancillary data about each band.  A value of type `Band`
is now guaranteed to represent a _single_ band.  Functions that
operate on one or both bands now take a list-of-Band argument instead
of a `Band` argument.