glcat/cataloging.py: refactor out the “data” dictionary
`accumulate_run_data` is removed.  `make_catalog` is renamed
`make_band_catalog`; it reads and writes almost all the data it needs
itself, and does not read any data it doesn’t need.

There’s two exceptions to “almost all the data it needs.”  Exposure
times are loaded up by a new function `exposure_times_for_catalog`.
This function is also responsible for crunching the exposure time
tables down to the two numbers that `make_band_catalog` wants.
The observation type is extracted from eclipse metadata in the
higher-level function `band_catalog`, which was already responsible
for dealing with multi-leg eclipses.

`exposure_times_for_catalog` also experiments a little with avoiding
use of Pandas.  Instead we work directly with PyArrow.  We may not be
doing so optimally but it doesn’t really matter because the exposure
time tables are relatively short.