glcat/cataloging.py: don’t use pandas
Work directly with pyarrow instead.

Validate the assumption that all of the per-aperture input tables
have the same source index (i.e. the ra and dec columns), and sort
these tables by that index to ensure that the big cross-aperture
join lines up correctly.

This is good for about 10% speedup despite the cost of the sorting and
validation.  This stage of processing is not the bottleneck, but on a
data set of GALEX’s size every little bit helps.

Type annotations may not be quite right.